+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@dimo-network/dimo-node-sdk-dev",
-  "version": "1.2.01",
+  "version": "1.2.21",
   "description": "DIMO SDK",
   "keywords": [
     "DIMO"
