+++ src/index.ts
@@ -16,4 +16,4 @@ export {
 } from "./core/actions/claimAndPairAftermarketDevice.js";
 export { kernelClientFromPasskeySigner } from "./core/kernelClientFromSigner/kernelClientFromPasskey.js";
 export { kernelClientFromPrivateKey } from "./core/kernelClientFromSigner/kernelClientFromPrivateKey.js";
-export { KernelSigner } from "./DIMODeveloperSDK.js";
+export { KernelSigner } from "./KernelSigner.js";
