+++ src/version.ts
@@ -1,1 +1,1 @@
-export const LIB_VERSION = '1.2.01';
+export const LIB_VERSION = '1.2.21';
